requirements.txt
streamlit>=1.47
groq>=0.30
pandas>=2.3
matplotlib>=3.10
sqlite3-bro
python-dotenv>=1.0